CRUD USER Corregir mensajes de salida

+++ app/Controllers/User.php
@@ -12,18 +12,16 @@ class User extends ResourceController {
     protected $format    = 'json';
 
     public function index() {
-        return $this->genericResponse($this->model->findAll(), "", 200);
+        $rows = $this->model->findAll();
+        return $this->genericResponse($rows, ["total" => count($rows)], null, 200);
     }
 
     public function show($id = null) {
-        if($id == null){
-            return $this->genericResponse(null, "ID es requerido", 400);
-        }
         $user = $this->model->find($id);
         if(!$user){
-            return $this->genericResponse(null, "User no existe", 404);
+            return $this->genericResponse(null, "Registro inexistente", "No existe el registro $id", 404);
         }
-        return $this->genericResponse($user, "", 200);
+        return $this->genericResponse($user, "Se ha encontrado el registro", null, 200);
     }
 
     public function create(){
@@ -32,10 +30,10 @@ class User extends ResourceController {
 
         if($this->validate('user_create')){
             if(!$this->request->getPost('id_rol')){
-                return $this->genericResponse(null, array('id_rol' => 'Role es requerido'), 400);
+                return $this->genericResponse(null, "Error de validación", array('id_rol' => 'Rol es requerido'), 400);
             }
             if(!$role->get($this->request->getPost('id_rol'))){
-                return $this->genericResponse(null, array('id_rol' => 'Role no existe'), 404);
+                return $this->genericResponse(null, "Registro inexistente", array('id_rol' => 'Rol no existe'), 404);
             }
             $id = $user->insert([
                 'username' => $this->request->getPost('username'),
@@ -43,10 +41,10 @@ class User extends ResourceController {
                 'email' => $this->request->getPost('email'),
                 'id_rol' => $this->request->getPost('id_rol'),
             ]);
-            return $this->genericResponse($this->model->find($id),null,200);
+            return $this->genericResponse($this->model->find($id), "Registro adicionado correctamente", null, 200);
         }
         $validation = \Config\Services::validation();
-        return $this->genericResponse(null, $validation->getErrors(), 400);
+        return $this->genericResponse(null, "Error de validación", $validation->getErrors(), 400);
     }
 
     public function update($id = null){
@@ -54,17 +52,17 @@ class User extends ResourceController {
         $role = new RoleModel();
 
         if(!$this->model->find($id)){
-            return $this->genericResponse(null, "User no existe", 404);
+            return $this->genericResponse(null, "Registro inexistente", "User no existe", 404);
         }
 
         $data = $this->request->getRawInput();
 
         if($this->validate('user_update')){
-            if(!$data['id_rol']){
-                return $this->genericResponse(null, array('id_rol' => 'Role es requerido'), 400);
+            if(!isset($data['id_rol'])){
+                return $this->genericResponse(null, "Error de validación", array('id_rol' => 'Rol es requerido'), 400);
             }
             if(!$role->get($data['id_rol'])){
-                return $this->genericResponse(null, array('id_rol' => 'Role no existe'), 404);
+                return $this->genericResponse(null, "Registro inexistente", array('id_rol' => 'Rol no existe'), 404);
             }
             var_dump($data['username']);
             $user->update($id, [
@@ -72,27 +70,32 @@ class User extends ResourceController {
                 'email' => $data['email'],
                 'id_rol' => $data['id_rol']
             ]);
-            return $this->genericResponse($this->model->find($id),null,200);
+            return $this->genericResponse($this->model->find($id), "Registro $id editado correctamente", null,200);
         }
         $validation = \Config\Services::validation();
-        return $this->genericResponse(null, $validation->getErrors(), 400);
+        return $this->genericResponse(null, "Error de validación", $validation->getErrors(), 400);
     }
 
     public function delete($id = null) {
+        if(!$this->model->find($id)){
+            return $this->genericResponse(null, "Registro inexistente", "User no existe", 404);
+        }
         $this->model->delete($id);
-        return $this->genericResponse("Registro $id eliminado correctamente", null, 200);
+        return $this->genericResponse(null, "Registro $id eliminado correctamente", null, 200);
     }
     
-    private function genericResponse($data, $msg, $code){
+    private function genericResponse($data, $message, $error, $code){
         if($code == 200){
             return $this->respond(array(
+                "ok" => true,
+                "message" => $message,
                 "data" => $data,
-                "code" => $code
             ));
-        }else{
+        } else {
             return $this->respond(array(
-                "msg" => $msg,
-                "code" => $code
+                "ok" => false,
+                "message" => $message,
+                "error" => $error,                
             ), $code);
         }
     }
